fix: fix TESFile::Duplicate hook
The hook set the smallCompileIndex after the duplication was complete, but the smallCompileIndex is used in SE during the duplication process.

The hook now sets the index in the exact spot SE does to ensure same behavior

Additionally, refactored `IndexCheck` to move some logic away from ASM to C++ to be on the safer side.